requirements.txt
python-docx
googletrans==4.0.0-rc1
